+++ aauth-axum/src/person/mod.rs
@@ -3,6 +3,6 @@ mod routes;
 pub use routes::{
     PersonServerState, interaction_callback_handler, interaction_start_handler,
     pending_clarification_post_handler, pending_poll_handler, pending_post_handler,
-    person_jwks_handler, person_metadata_handler, token_exchange_deferred_handler,
+    person_jwks_handler, person_metadata_handler, person_router, token_exchange_deferred_handler,
     token_exchange_handler,
 };