+++ aauth-axum/src/resource/mod.rs
@@ -4,4 +4,4 @@ mod routes;
 
 pub use extract::VerifiedAAuthToken;
 pub use layer::ResourceAuthLayer;
-pub use routes::{ResourceServerState, resource_pending_poll_handler};
+pub use routes::{ResourceServerState, resource_pending_poll_handler, resource_router};